refactor(plugin): extract ShadowStartupCoordinator from main.ts (0.4.70)
PR 2/3 of the main.ts god-file split. Lifts the plugin-install half
of shadow-vault startup into a single owner class.

Moved (4 private methods + 1 helper, ~150 LOC of method bodies):
- handlePendingPluginSuggestions
- installMissingShadowPlugins
- copyPluginConfigsForInstalled
- makeMarketplaceInstaller (private helper of the above)
- runShadowStartup body shrunk to "construct coordinator → call
  prepareForAutoConnect() → runAutoConnect"

Coordinator boundary:
- Constructor takes (app, settings, saveSettings callback)
- Public API: prepareForAutoConnect() — runs the two install passes in order
- Owns the lifecycle of settings.pendingPluginSuggestions: clears the
  field and calls saveSettings on a definitive install/skip decision

Behavior is byte-identical: same modal flow, same logger lines, same
notice strings, same data.json paths and JSON formatting (2-space +
trailing newline). The FileSystemAdapter probe and getBasePath calls
are now in the coordinator (where they belong) instead of main.ts.

main.ts: 1641 → 1478 LOC (−163). One less concern, three fewer
imports, the file's narrative is closer to "plugin lifecycle".

569/569 tests pass, typecheck clean.
