+++ plugin/versions.json
@@ -73,5 +73,6 @@
   "0.4.66": "1.4.0",
   "0.4.67": "1.4.0",
   "0.4.68": "1.4.0",
-  "0.4.69": "1.4.0"
+  "0.4.69": "1.4.0",
+  "0.4.70": "1.4.0"
 }
